+++ NEWS.md
@@ -1,4 +1,13 @@
-## qtl2geno 0.5-4
+## qtl2geno 0.5-5 (2017-03-17)
+
+### New features
+
+- Added a function `find_ibd_segments` that takes genotypes for a set
+  of inbred strains and searches for segments where strain pairs look
+  to be IBD.
+
+
+## qtl2geno 0.5-4 (2017-03-13)
 
 ### New features